Refactor chat search logic from ViewModel to Repository
Moves the search implementation, including precomputed chat indices and caching, out of `ChatViewModel` into `ChatRepositoryImpl`. Introduces a domain-level `ChatSearchMode` enum to avoid leaking presentation layer search toggles to the repository. The ViewModel now safely delegates queries directly.
