+++ app/src/main/java/org/ole/planet/myplanet/repository/ChatRepositoryImpl.kt
@@ -22,8 +22,11 @@ import org.ole.planet.myplanet.model.Conversation
 import org.ole.planet.myplanet.model.Data
 import org.ole.planet.myplanet.model.News
 import org.ole.planet.myplanet.services.SharedPrefManager
+import kotlinx.coroutines.withContext
 import org.ole.planet.myplanet.services.sync.ServerUrlMapper
+import org.ole.planet.myplanet.utils.DispatcherProvider
 import org.ole.planet.myplanet.utils.JsonUtils
+import org.ole.planet.myplanet.utils.Utilities
 
 @Singleton
 class ChatRepositoryImpl @Inject constructor(
@@ -31,8 +34,19 @@ class ChatRepositoryImpl @Inject constructor(
     private val chatApiService: ChatApiService,
     private val serverUrlMapper: ServerUrlMapper,
     private val sharedPrefManager: SharedPrefManager,
+    private val dispatcherProvider: DispatcherProvider,
     @PlainGson private val gson: Gson
 ) : ChatRepository, ChatSyncWriter {
+
+    private data class PrecomputedChat(
+        val chat: ChatHistory,
+        val normalizedTitle: String?,
+        val normalizedQueries: List<String?>,
+        val normalizedResponses: List<String?>
+    )
+
+    private var allChats: List<ChatHistory> = emptyList()
+    private var precomputedChats: List<PrecomputedChat> = emptyList()
 
     @VisibleForTesting
     internal var reachabilityCheck: suspend (String) -> Boolean = { url ->
@@ -118,9 +132,92 @@ class ChatRepositoryImpl @Inject constructor(
 
     override suspend fun getChatHistoryForUser(userName: String?): List<ChatHistory> {
         if (userName.isNullOrEmpty()) {
+            allChats = emptyList()
+            precomputedChats = emptyList()
             return emptyList()
         }
-        return chatDao.getByUser(userName)
+        val chats = chatDao.getByUser(userName)
+        allChats = sortChats(chats)
+        precomputedChats = buildPrecomputedChats(allChats)
+        return allChats
+    }
+
+    private fun sortChats(chats: List<ChatHistory>): List<ChatHistory> {
+        return chats.sortedByDescending { chat ->
+            maxOf(chat.createdDate?.toLongOrNull() ?: 0L, chat.updatedDate?.toLongOrNull() ?: 0L)
+        }
+    }
+
+    private suspend fun buildPrecomputedChats(chats: List<ChatHistory>): List<PrecomputedChat> = withContext(dispatcherProvider.default) {
+        chats.map { chat ->
+            val title = if (chat.conversations != null && chat.conversations?.isNotEmpty() == true) {
+                chat.conversations?.get(0)?.query?.let { Utilities.normalizeText(it) }
+            } else {
+                chat.title?.let { Utilities.normalizeText(it) }
+            }
+            val queries = chat.conversations?.map { it?.query?.let { q -> Utilities.normalizeText(q) } } ?: emptyList()
+            val responses = chat.conversations?.map { it?.response?.let { r -> Utilities.normalizeText(r) } } ?: emptyList()
+            PrecomputedChat(chat, title, queries, responses)
+        }
+    }
+
+    override suspend fun searchChats(query: String, mode: ChatSearchMode): List<ChatHistory> {
+        return if (mode == ChatSearchMode.TITLE) {
+            searchByTitle(query)
+        } else {
+            fullConvoSearch(query, isQuestion = (mode == ChatSearchMode.QUESTION))
+        }
+    }
+
+    private suspend fun fullConvoSearch(s: String, isQuestion: Boolean): List<ChatHistory> = withContext(dispatcherProvider.default) {
+        var conversation: String?
+        val queryParts = s.split(" ").filterNot { it.isEmpty() }
+        val normalizedQueryParts = queryParts.map { Utilities.normalizeText(it) }
+        val normalizedQuery = Utilities.normalizeText(s)
+        val inTitleStartQuery = mutableListOf<ChatHistory>()
+        val inTitleContainsQuery = mutableListOf<ChatHistory>()
+        val startsWithQuery = mutableListOf<ChatHistory>()
+        val containsQuery = mutableListOf<ChatHistory>()
+        for (pChat in precomputedChats) {
+            val conversations = pChat.chat.conversations
+            if (!conversations.isNullOrEmpty()) {
+                for (i in 0 until conversations.size) {
+                    conversation = if (isQuestion) {
+                        pChat.normalizedQueries[i]
+                    } else {
+                        pChat.normalizedResponses[i]
+                    }
+                    if (conversation == null) continue
+                    if (conversation.startsWith(normalizedQuery, ignoreCase = true)) {
+                        if (i == 0) inTitleStartQuery.add(pChat.chat) else startsWithQuery.add(pChat.chat)
+                        break
+                    } else if (normalizedQueryParts.all { conversation.contains(it, ignoreCase = true) }) {
+                        if (i == 0) inTitleContainsQuery.add(pChat.chat) else containsQuery.add(pChat.chat)
+                        break
+                    }
+                }
+            }
+        }
+        inTitleStartQuery + inTitleContainsQuery + startsWithQuery + containsQuery
+    }
+
+    private suspend fun searchByTitle(s: String): List<ChatHistory> = withContext(dispatcherProvider.default) {
+        var title: String?
+        val queryParts = s.split(" ").filterNot { it.isEmpty() }
+        val normalizedQueryParts = queryParts.map { Utilities.normalizeText(it) }
+        val normalizedQuery = Utilities.normalizeText(s)
+        val startsWithQuery = mutableListOf<ChatHistory>()
+        val containsQuery = mutableListOf<ChatHistory>()
+        for (pChat in precomputedChats) {
+            title = pChat.normalizedTitle
+            if (title == null) continue
+            if (title.startsWith(normalizedQuery, ignoreCase = true)) {
+                startsWithQuery.add(pChat.chat)
+            } else if (normalizedQueryParts.all { title.contains(it, ignoreCase = true) }) {
+                containsQuery.add(pChat.chat)
+            }
+        }
+        startsWithQuery + containsQuery
     }
 
     override suspend fun getLatestRev(id: String): String? {
