+++ app/src/test/java/org/ole/planet/myplanet/ui/chat/ChatViewModelTest.kt
@@ -202,51 +202,12 @@ class ChatViewModelTest {
     }
 
     @Test
-    fun `searchChats by title correctly filters list`() = runTest {
-        val chat1 = ChatHistory().apply { title = "First Chat" }
-        val chat2 = ChatHistory().apply { title = "Second Discussion" }
-
-        coEvery { chatRepository.getChatHistoryForUser(any()) } returns listOf(chat1, chat2)
-
-        viewModel.loadChatHistoryScreenData("user123", null, null)
-        testScheduler.advanceUntilIdle()
-
-        viewModel.searchChats("First", isFullSearch = false, isQuestion = false)
-        testScheduler.advanceUntilIdle()
-
-        assertEquals(1, viewModel.filteredChats.value.size)
-        assertEquals("First Chat", viewModel.filteredChats.value[0].title)
-    }
-
-    @Test
-    fun `searchChats by full conversation filters by question`() = runTest {
-        val chat1 = ChatHistory().apply {
-            title = "Chat 1"
-            conversations = listOf(Conversation().apply { query = "How is the weather?" })
-        }
-        val chat2 = ChatHistory().apply {
-            title = "Chat 2"
-            conversations = listOf(Conversation().apply { query = "Tell me a joke." })
-        }
-
-        coEvery { chatRepository.getChatHistoryForUser(any()) } returns listOf(chat1, chat2)
-
-        viewModel.loadChatHistoryScreenData("user123", null, null)
-        testScheduler.advanceUntilIdle()
-
-        viewModel.searchChats("weather", isFullSearch = true, isQuestion = true)
-        testScheduler.advanceUntilIdle()
-
-        assertEquals(1, viewModel.filteredChats.value.size)
-        assertEquals("Chat 1", viewModel.filteredChats.value[0].title)
-    }
-
-    @Test
-    fun `searchChats with empty query resets filtered list`() = runTest {
+    fun `searchChats delegates to repository correctly and empty query resets filtered list`() = runTest {
         val chat1 = ChatHistory().apply { title = "Chat 1" }
         val chat2 = ChatHistory().apply { title = "Chat 2" }
 
         coEvery { chatRepository.getChatHistoryForUser(any()) } returns listOf(chat1, chat2)
+        coEvery { chatRepository.searchChats("Chat 1", org.ole.planet.myplanet.repository.ChatSearchMode.TITLE) } returns listOf(chat1)
 
         viewModel.loadChatHistoryScreenData("user123", null, null)
         testScheduler.advanceUntilIdle()
@@ -258,6 +219,8 @@ class ChatViewModelTest {
         viewModel.searchChats("", isFullSearch = false, isQuestion = false)
         testScheduler.advanceUntilIdle()
         assertEquals(2, viewModel.filteredChats.value.size)
+
+        coVerify { chatRepository.searchChats("Chat 1", org.ole.planet.myplanet.repository.ChatSearchMode.TITLE) }
     }
 
     @Test
